Use common action for web and mobile to fetch user supporter
