+++ packages/mobile/src/message/types.ts
@@ -128,8 +128,7 @@ export enum MessageType {
 
   // Tipping
   FETCH_RECENT_TIPS = 'fetch-recent-tips',
-  UPDATE_TIPS_STORAGE = 'update-tips-storage',
-  FETCH_USER_SUPPORTER = 'fetch-user-supporter'
+  UPDATE_TIPS_STORAGE = 'update-tips-storage'
 }
 
 export type Message = {
