+++ packages/web/src/common/store/tipping/slice.ts
@@ -149,6 +149,14 @@ const slice = createSlice({
     ) => {
       state.recentTips = action.payload.recentTips
     },
+    fetchUserSupporter: (
+      state,
+      action: PayloadAction<{
+        currentUserId: ID
+        userId: ID
+        supporterUserId: ID
+      }>
+    ) => {},
     setTipToDisplay: (
       state,
       action: PayloadAction<{ tipToDisplay: UserTip }>
@@ -177,6 +185,7 @@ export const {
   resetSend,
   fetchRecentTips,
   setRecentTips,
+  fetchUserSupporter,
   setTipToDisplay,
   hideTip
 } = slice.actions
