+++ packages/web/src/services/native-mobile-interface/types.ts
@@ -133,8 +133,7 @@ export enum MessageType {
 
   // Tipping
   FETCH_RECENT_TIPS = 'fetch-recent-tips',
-  UPDATE_TIPS_STORAGE = 'update-tips-storage',
-  FETCH_USER_SUPPORTER = 'fetch-user-supporter'
+  UPDATE_TIPS_STORAGE = 'update-tips-storage'
 }
 
 export interface Message {
